Overlayed arrowheads over the draggable circles
