+++ content/complex/components/complexarithmetic.less
@@ -1,5 +1,5 @@
 // =============================================================================
-// ComplexTransform Component
+// ComplexArithmetic Component
 // (c) Mathigon
 // =============================================================================
 
@@ -7,7 +7,7 @@
 @import "../../shared/variables";
 
 
-x-complex-transform {
+x-complex-arithmetic {
   display: block;
   position: relative;
   margin: 1.5rem 0;
