+++ content/complex/components/complexarithmetic.pug
@@ -1,4 +1,4 @@
-template#complex-transform
+template#complex-arithmetic
   x-select.tabs
     for i in ["Sum","Product","Difference","Quotient"]
       .group(data-value=i)= i
