+++ content/complex/styles.less
@@ -1,3 +1,3 @@
 @import "../shared/variables";
-@import "./components/complextransform";
+@import "./components/complexarithmetic";
